Добавлена domain модель  вакансии в пакет search изменено название ответа от сервера на VacancyDto, определена сущность БД
