+++ app/src/main/java/ru/practicum/android/diploma/common/model/VacancyDto.kt
