--- app/src/main/java/ru/practicum/android/diploma/data/AppDatabase.kt
@@ -1,18 +1,0 @@
-package ru.practicum.android.diploma.data
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-import ru.practicum.android.diploma.data.dao.FavoriteVacancyDao
-import ru.practicum.android.diploma.data.dao.VacancyDao
-import ru.practicum.android.diploma.data.entity.FavoriteVacancyEntity
-import ru.practicum.android.diploma.data.entity.VacancyEntity
-
-@Database(
-    entities = [VacancyEntity::class, FavoriteVacancyEntity::class],
-    version = 1,
-    exportSchema = false
-)
-abstract class AppDatabase : RoomDatabase() {
-    abstract fun vacancyDao(): VacancyDao
-    abstract fun favoriteVacancyDao(): FavoriteVacancyDao
-}
