--- app/src/main/java/ru/practicum/android/diploma/data/entity/FavoriteVacancyEntity.kt
@@ -1,9 +1,0 @@
-package ru.practicum.android.diploma.data.entity
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "favorites")
-data class FavoriteVacancyEntity(
-    @PrimaryKey val id: String
-)
