+++ app/src/main/java/ru/practicum/android/diploma/favorites/data/db/AppDatabase.kt
@@ -1,0 +1,13 @@
+package ru.practicum.android.diploma.favorites.data.db
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+
+@Database(
+    entities = [VacancyEntity::class],
+    version = 1,
+    exportSchema = false
+)
+abstract class AppDatabase : RoomDatabase() {
+    abstract fun vacancyDao(): VacancyDao
+}
